--- a/.changeset/tall-owls-lie.md
+++ b/.changeset/tall-owls-lie.md
@@ -1,0 +1,5 @@
+---
+"@chakra-ui/storybook-addon": patch
+---
+
+Update storybook dependency