--- a/.storybook/main.js
+++ b/.storybook/main.js
@@ -1,6 +1,12 @@
 module.exports = {
   core: {
-    builder: "webpack5",
+    builder: {
+      name: "webpack5",
+      options: {
+        lazyCompilation: true,
+        fsCache: true,
+      },
+    },
   },
   stories: ["../packages/**/stories/*.stories.tsx"],
   addons: [