--- a/tooling/storybook-addon/package.json
+++ b/tooling/storybook-addon/package.json
@@ -34,9 +34,9 @@
   "homepage": "https://github.com/chakra-ui/chakra-ui#readme",
   "devDependencies": {
     "@chakra-ui/react": "^2.1.2",
-    "@storybook/addons": "^6.4.22",
-    "@storybook/api": "~6.5.0-alpha.64",
-    "@storybook/components": "~6.5.0-alpha.64",
+    "@storybook/addons": "^6.5.6",
+    "@storybook/api": "^6.5.6",
+    "@storybook/components": "^6.5.6",
     "concurrently": "^6.4.0",
     "react": "^18.0.0"
   },