<!---
Thanks for creating a Pull Request 💖!

Please read the following before submitting:
- PRs that adds new external dependencies might take a while to review.
- Keep your PR as small as possible.
- Limit your PR to one type (docs, feature, refactoring, ci, or bugfix)
-->

## 📝 Description

Enable `lazyCompilation` and `fsCache` to speed up storybook builds

### Benchmarks
|  |  |
| --- | --- |
| Before | `24 s for manager and 26 s for preview` |
| After (w/o cache) | `23 s for manager and 24 s for preview` |
| After (w/ cache) | `11 s for manager and 9.37 s for preview` |

## 💣 Is this a breaking change (Yes/No):

No

## 📝 Additional Information